+++ assets/jsons/client_config.json
@@ -1,10 +1,4 @@
 {
-  "hotel_visit_purpose_probabilities": {
-    "VACATION": 0.7,
-    "BUSINESS_TRIP": 0.2,
-    "REHABILITATION": 0.1
-  },
-
   "desired_rank_probabilities": {
     "VACATION": {
       "ECONOMIC": 20,
