+++ assets/jsons/game_config.json
@@ -5,5 +5,12 @@
 
   "employees_to_hire_list_size": 12,
   "possible_employee_removal_probability": 0.05,
-  "room_fault_probability": 0.01
+  "room_fault_probability": 0.01,
+
+  "difficulty_multiplier": {
+    "EASY": 0.5,
+    "MEDIUM": 1.0,
+    "HARD": 1.5,
+    "BRUTAL": 2.0
+  }
 }
