+++ assets/jsons/hotel_scenarios_config.json
@@ -19,14 +19,6 @@
     }
   },
 
-
-  "difficulty_multiplier": {
-    "EASY": 0.5,
-    "MEDIUM": 1.0,
-    "HARD": 1.5,
-    "BRUTAL": 2.0
-  },
-
   "clients_per_month_multiplier": {
     "HOTEL": {"1": 0.8, "2": 0.85, "3": 0.9, "4": 1.2, "5": 1.3, "6": 1.1, "7": 0.8, "8": 0.7, "9": 0.9, "10": 1.1, "11": 1.0, "12": 0.5},
     "RESORT": {"1": 1.1, "2": 1.0, "3": 0.9, "4": 0.9, "5": 1.0, "6": 1.1, "7": 1.5, "8": 1.4, "9": 1.1, "10": 0.9, "11": 0.8, "12": 0.6},
